requirements.txt
# Core dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0

# COCO utilities
pycocotools>=2.0.6

# Utilities
requests>=2.31.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
